Moved some stuff to the greenhouse-effect scene view from the base-greenhouse scene view
