+++ greenhouse-effect/src/js/views/scene/base-greenhouse.js
@@ -9,7 +9,6 @@ define(function(require) {
     var ModelViewTransform = require('common/math/model-view-transform');
     var Rectangle          = require('common/math/rectangle');
     var Vector2            = require('common/math/vector2');
-    var range              = require('common/math/range');
 
     var PhotonView = require('views/photon');
     var CloudView  = require('views/cloud');
@@ -35,8 +34,6 @@ define(function(require) {
             this.listenTo(this.simulation.photons, 'reset',          this.photonsReset);
             this.listenTo(this.simulation.photons, 'add',            this.photonAdded);
             this.listenTo(this.simulation.photons, 'remove destroy', this.photonRemoved);
-
-            this.listenTo(this.simulation.atmosphere, 'change:greenhouseGasConcentration', this.updatePollution);
         },
 
         renderContent: function() {
@@ -55,7 +52,6 @@ define(function(require) {
             this.initMVT();
             this.initBackground();
             this.initPhotons();
-            this.initPolution();
 
             this.initialized = true;
         },
@@ -115,63 +111,8 @@ define(function(require) {
             this.photonsReset(this.simulation.photons);
         },
 
-        initPolution: function() {
-            if (this.pollution)
-                this.foregroundLayer.removeChild(this.pollution);
-
-            var canvas = document.createElement('canvas');
-            canvas.width  = this.width;
-            canvas.height = this.height;
-
-            var ctx = canvas.getContext('2d');
-
-            var gradient = ctx.createLinearGradient(0, 0, 0, this.height);
-            gradient.addColorStop(0, Constants.Atmosphere.POLLUTION_TOP_COLOR);
-            gradient.addColorStop(1, Constants.Atmosphere.POLLUTION_BOTTOM_COLOR);
-            ctx.fillStyle = gradient;
-            ctx.fillRect(0, 0, this.width, this.height);
-
-            this.pollution = new PIXI.Sprite(PIXI.Texture.fromCanvas(canvas));
-            this.foregroundLayer.addChild(this.pollution);
-
-            this.pollutionRange = range({
-                min: Constants.Atmosphere.MIN_GREENHOUSE_GAS_CONCENTRATION,
-                max: Constants.Atmosphere.MAX_GREENHOUSE_GAS_CONCENTRATION
-            });
-
-            this.updatePollution(this.simulation.atmosphere, this.simulation.atmosphere.get('greenhouseGasConcentration'));
-        },
-
-        resize: function() {
-            PixiSceneView.prototype.resize.apply(this, arguments);
-
-            if (this.initialized) {
-                this.setSceneScale(this.bgToday);
-                this.setSceneScale(this.bg1750);
-                this.setSceneScale(this.bgIceAge);
-                this.initPolution();
-            }
-        },
-
         _update: function(time, deltaTime, paused, timeScale) {
             
-        },
-
-        updatePollution: function(atmosphere, concentration) {
-            this.pollution.alpha = 0.2 * this.pollutionRange.percent(concentration);
-        },
-
-        photonsReset: function(photons) {
-            // Remove old photon views
-            for (var i = this.photonViews.length - 1; i >= 0; i--) {
-                this.photonViews[i].removeFrom(this.photons);
-                this.photonViews.splice(i, 1);
-            }
-
-            // Add new photon views
-            photons.each(function(photon) {
-                this.createAndAddPhotonView(photon);
-            }, this);
         },
 
         photonAdded: function(photon, photons) {
@@ -186,6 +127,19 @@ define(function(require) {
                     break;
                 }
             }
+        },
+
+        photonsReset: function(photons) {
+            // Remove old photon views
+            for (var i = this.photonViews.length - 1; i >= 0; i--) {
+                this.photonViews[i].removeFrom(this.photons);
+                this.photonViews.splice(i, 1);
+            }
+
+            // Add new photon views
+            photons.each(function(photon) {
+                this.createAndAddPhotonView(photon);
+            }, this);
         },
 
         createAndAddPhotonView: function(photon) {
